Glorious first version of Constant::Export::Lazy
What this is and why is explained in the POD. This is something I've
been meaning to write for a while but hadn't gotten around to it.

Some maybe TODOs include, I might never get around to these:

 * Being able to do `perl -MYour::Exporter=SOME_CONSTANT` to spew out
   the value of that constant.

 * Ability to generate constants that'll never be exported (maybe
   matching a regex like qr/^_/). Makes it possible to have
   convenience constants for making other constants that you'll never
   export anywhere.

Things I was once going to do, but decided not to:

 * Ability to pass a subref that will validate the constant (either
   from our code or the stuff coming from %ENV)

 * Spewing out diagnostics in some debugging mode where all constants
   are imported, so you can see what the full setup of the env
   is. Needs a "documentation" key or something.

These can both be done by just dying or doing a "print STDERR" in the
callback that defines constants, respectively.