Implement name munging and solve foreign subroutine edge cases
This patch that simultaneously adds a new feature and fixes a few bugs
was hard to split up, so I'm commiting this all in one go.

This solves and tests for cases where we tried to call
constant.pm-generated subroutines and would previously fail to correctly
unpack their values from the stash (since constant.pm cheats and doesn't
actually create subs).

This also implements an unrelated feature where we can now munge the
internal names we have for constants we're generating. See the
documentation for what this is for.