Simplify the lookup of values to not look directly at the stash
So the internals of perl will create a $symtab->{$whatever} whose ref()
returns SCALAR or REF or whatever when constant.pm creates constants.

But rather than deal with that complexity just use another syntax that
does this all for us.

This simplifies our code a lot as we don't have to deal with the
internals of how perl core itself is managing these values.